Break transaction* structs into separate file
